# src/app/api/openapi.yaml
openapi: 3.0.0
info:
  title: ZenZen CRM API
  version: '1.0.0'
  description: |
    API for the ZenZen CRM system, providing ticket management, messaging, and team collaboration features.
  contact:
    name: ZenZen Support
    url: https://zenzen.support
servers:
  - url: https://api.zenzen.dev
    description: Development server
  - url: https://api.zenzen.com
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [UNOPENED, IN PROGRESS, RESOLVED, UNRESOLVED]
        priority:
          type: string
          enum: [NONE, LOW, MEDIUM, HIGH, CRITICAL]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        first_response_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
        reopen_count:
          type: integer
        assigned_to:
          type: string
          format: uuid
        assigned_team:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
        custom_fields:
          type: object
          additionalProperties: true
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ticket_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
...
    File:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ticket_id:
          type: string
          format: uuid
        file_url:
          type: string
        uploaded_at:
          type: string
          format: date-time
...
paths:
...
  /tickets/{ticket_id}/files:
    parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Upload a file to a ticket
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
...
  /search:
    get:
      summary: Search across tickets and messages
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [tickets, messages, all]
            default: all
        - name: status
          in: query
          schema:
            type: string
            enum: [UNOPENED, IN PROGRESS, RESOLVED, UNRESOLVED]
        - name: priority
          in: query
          schema:
            type: string
            enum: [NONE, LOW, MEDIUM, HIGH, CRITICAL]
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ticket'
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      pages:
                        type: integer
                      current_page:
                        type: integer
                      per_page:
                        type: integer
...
security:
  - ApiKeyAuth: []
  - BearerAuth: [] 